# Optimize load times with JIT on-demand