… galois/field/meta_gfpm.py
… galois/field/meta_ufunc.py
… galois/group/meta_ufunc_add.py
… galois/group/meta_ufunc_mul.py