Don't compile all ufuncs and JIT functions at class build time, but at first-use time.

This prevents users from paying a load penalty for ufuncs or functions they never use. For those ufuncs/functions they do use, there is a slight delay (~100 ms) in the first call to JIT compile and then subsequent calls are fast. Even if all ufuncs/functions are used, this distributes the compile/load time.

Before:
```python
In [1]: %time import galois
CPU times: user 1.71 s, sys: 688 ms, total: 2.4 s
Wall time: 1.25 s

In [2]: %time GF = galois.GF(2**8)
CPU times: user 935 ms, sys: 11.6 ms, total: 946 ms
Wall time: 953 ms

In [3]: a = GF.Random(10)

In [4]: b = GF.Random(10)

In [5]: %time a / b
CPU times: user 544 µs, sys: 123 µs, total: 667 µs
Wall time: 629 µs
Out[5]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)

In [6]: %time a / b
CPU times: user 332 µs, sys: 75 µs, total: 407 µs
Wall time: 344 µs
Out[6]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)

In [7]: %time a / b
CPU times: user 260 µs, sys: 59 µs, total: 319 µs
Wall time: 217 µs
Out[7]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)
```

After:
```python
In [1]: %time import galois                                                                     
CPU times: user 1.22 s, sys: 702 ms, total: 1.92 s
Wall time: 684 ms

In [2]: %time GF = galois.GF(2**8)                                                              
CPU times: user 239 ms, sys: 683 µs, total: 240 ms
Wall time: 239 ms

In [3]: a = GF.Random(10)                                                                       

In [4]: b = GF.Random(10, low=1)                                                                

In [5]: %time a / b                                                                             
CPU times: user 96.2 ms, sys: 0 ns, total: 96.2 ms
Wall time: 95 ms
Out[5]: GF([  3, 129, 103, 253, 116, 198, 173, 109, 234, 165], order=2^8)

In [6]: %time a / b                                                                             
CPU times: user 407 µs, sys: 128 µs, total: 535 µs
Wall time: 462 µs
Out[6]: GF([  3, 129, 103, 253, 116, 198, 173, 109, 234, 165], order=2^8)

In [7]: %time a / b                                                                             
CPU times: user 474 µs, sys: 146 µs, total: 620 µs
Wall time: 552 µs
Out[7]: GF([  3, 129, 103, 253, 116, 198, 173, 109, 234, 165], order=2^8)
```