@mhostetter 2021-06-03
Before:
```python
In [1]: %time import galois
CPU times: user 1.71 s, sys: 688 ms, total: 2.4 s
Wall time: 1.25 s

In [2]: %time GF = galois.GF(2**8)
CPU times: user 935 ms, sys: 11.6 ms, total: 946 ms
Wall time: 953 ms

In [3]: a = GF.Random(10)

In [4]: b = GF.Random(10)

In [5]: %time a / b
CPU times: user 544 µs, sys: 123 µs, total: 667 µs
Wall time: 629 µs
Out[5]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)

In [6]: %time a / b
CPU times: user 332 µs, sys: 75 µs, total: 407 µs
Wall time: 344 µs
Out[6]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)

In [7]: %time a / b
CPU times: user 260 µs, sys: 59 µs, total: 319 µs
Wall time: 217 µs
Out[7]: GF([132, 116,  38, 245,  71,  63,  23,  58, 247, 159], order=2^8)
```

After:
```python
In [1]: %time import galois
CPU times: user 1.25 s, sys: 722 ms, total: 1.97 s
Wall time: 725 ms

In [2]: %time GF = galois.GF(2**8)
CPU times: user 250 ms, sys: 0 ns, total: 250 ms
Wall time: 249 ms

In [3]: a = GF.Random(10)

In [4]: b = GF.Random(10, low=1)

In [5]: %time a / b
CPU times: user 121 ms, sys: 7.09 ms, total: 128 ms
Wall time: 127 ms
Out[5]: GF([ 92,  39,  96,  47, 192, 165, 197, 231,  74, 233], order=2^8)

In [6]: %time a / b
CPU times: user 141 µs, sys: 43 µs, total: 184 µs
Wall time: 169 µs
Out[6]: GF([ 92,  39,  96,  47, 192, 165, 197, 231,  74, 233], order=2^8)

In [7]: %time a / b
CPU times: user 0 ns, sys: 804 µs, total: 804 µs
Wall time: 763 µs
Out[7]: GF([ 92,  39,  96,  47, 192, 165, 197, 231,  74, 233], order=2^8)
```